compat_resource is no longer supported
this is breaking our internal use of the audit cookbook on chef-13
somehow by pulling this in as a dep.

this declares it drops support for chef <= 12.19 because that's the
last rev of chef-client that was pulled into compat_resource, i
don't know which features might specifically be getting used in
order to produce a lower threshold.  12.20.x is over a year old
though and that is still a wider window than our declared 6
month support window for cookbooks (and the EOL of the entire
Chef 12 line is imminently happening).

https://github.com/chef/chef-rfc/blob/master/rfc092-dependency-update-cadence.md#cookbook-and-ecosystem-tooling-support

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>