use Chef Server header auth in Chef Server fetcher
This function was removed in #383, likely because the original comment
on the function indicated that this was related to downloaded profiles
specifically from compliance.

In reality, this function overrides a function in the inspec Url
fetcher.  It overrides it so that it can use the Chef::ServerAPI
client. This client will use Chef Server's header based
authentication, allowing chef-clients to report audit data with a
specific data collector token.

Signed-off-by: Steven Danna <steve@chef.io>